requirements.txt
Django>=4.2
twilio
Pillow
reportlab
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
channels>=4.0
channels-redis>=4.1
daphne>=4.0
